Updated Projections to pan using Rhumb destinations instead of GreatCircle, the latter was causing gradual shift to equator.  Added Generics to Quadtree/DataOrganizer, refactored com.bbn.openmap.layer.location package, and csv subpackage.  db refactoring coming.  This refactoring makes it easier to extend by overriding single method.  Added Map<String, GraphicAttributes> to AbstractLocationHandler so entries can be set with different rendering attributes.  Updated docs for SimpleMap apps.

+++ src/openmap/com/bbn/openmap/layer/link/LinkLayer.java
@@ -1264,10 +1264,10 @@ public class LinkLayer extends OMGraphicHandlerLayer implements
         if (latmin >= -90.f && latmax <= 90.f && lonmin >= -180.f
                 && lonmax <= 180.f && latmin <= latmax && lonmin <= lonmax) {
             // Calculate center point
-            float dist = 0.5f * GreatCircle.sphericalDistance(ProjMath
+            double dist = 0.5f * GreatCircle.sphericalDistance(ProjMath
                     .degToRad(latmax), ProjMath.degToRad(lonmin), ProjMath
                     .degToRad(latmin), ProjMath.degToRad(lonmax));
-            float azimuth = GreatCircle.sphericalAzimuth(ProjMath
+            double azimuth = GreatCircle.sphericalAzimuth(ProjMath
                     .degToRad(latmax), ProjMath.degToRad(lonmin), ProjMath
                     .degToRad(latmin), ProjMath.degToRad(lonmax));
             center = GreatCircle.sphericalBetween(ProjMath.degToRad(latmax),
